fix: Refactored list components to work from a base class to increase reusability.
