+++ src/Web/kathanika-app/src/app/features/publications/pages/publication-list/publication-list.component.spec.ts
@@ -44,69 +44,4 @@ describe('PublicationListComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
-
-  it('hould change _queryVariables on changePage', () => {
-    let queryVariables = component.queryVariables;
-    const page = 2;
-
-    component.changePage(page);
-    expect(queryVariables.skip).toEqual((page - 1) * queryVariables.take);
-  });
-
-  it('should refetch query changePage', () => {
-    const queryRefRefetchSpy = spyOn<any>(component.queryRef, 'refetch');
-    const page = 2;
-
-    component.changePage(page);
-    expect(queryRefRefetchSpy).toHaveBeenCalled();
-  });
-
-  it('should change route query params on changePage', () => {
-    const page = 2;
-    component.changePage(page);
-    const routeQueryParams = router.getCurrentNavigation()?.extras
-      .queryParams as Params;
-    expect(routeQueryParams).toBeTruthy();
-    expect(routeQueryParams['page']).toEqual(page);
-  });
-
-  it('should change _queryVariables on changePageSize', () => {
-    let queryVariables = component.queryVariables;
-    const pageSize = 2;
-
-    component.changePageSize(pageSize);
-    expect(queryVariables.take).toEqual(pageSize);
-  });
-
-  it('should refetch query on changePageSize', () => {
-    const queryRefRefetchSpy = spyOn<any>(component.queryRef, 'refetch');
-    const pageSize = 2;
-
-    component.changePageSize(pageSize);
-    expect(queryRefRefetchSpy).toHaveBeenCalled();
-  });
-
-  it('should change route query params on changePageSize', () => {
-    const pageSize = 2;
-    component.changePageSize(pageSize);
-    const routeQueryParams = router.getCurrentNavigation()?.extras
-      .queryParams as Params;
-    expect(routeQueryParams).toBeTruthy();
-    expect(routeQueryParams['size']).toEqual(pageSize);
-  });
-
-  it('should change _QueryVariables on ngOnInit', () => {
-    // size: 3, page: 2
-    const queryVariables = component.queryVariables;
-    component.ngOnInit();
-    expect(queryVariables.skip).toEqual(20);
-    expect(queryVariables.take).toEqual(20);
-  });
-
-  it('should refetch query on ngOnInit', () => {
-    const queryVariables = component.queryVariables;
-    const queryRefRefetchSpy = spyOn<any>(component.queryRef, 'refetch');
-    component.ngOnInit();
-    expect(queryRefRefetchSpy).toHaveBeenCalledOnceWith(queryVariables);
-  });
 });
