+++ src/Web/kathanika-app/src/app/features/publications/pages/publication-list/publication-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Params, Router } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { MessageAlertService } from 'src/app/core/services/message-alert.service';
 import {
   GetPublicationsGQL,
@@ -7,20 +7,20 @@ import {
   GetPublicationsQueryVariables,
   SortEnumType,
 } from 'src/app/graphql/generated/graphql-operations';
-import { BaseQueryComponent } from 'src/app/shared/bases/base-query-component';
+import { BasePaginatedListComponent } from 'src/app/shared/bases/base-paginated-list-component';
 
 @Component({
   templateUrl: './publication-list.component.html',
   styleUrls: ['./publication-list.component.scss'],
 })
 export class PublicationListComponent
-  extends BaseQueryComponent<
+  extends BasePaginatedListComponent<
     GetPublicationsQuery,
     GetPublicationsQueryVariables
   >
-  implements OnInit
-{
-  private setQueryFilter(searchText: string) {
+  implements OnInit {
+
+  protected setSearchTextQueryFilter(searchText: string): void {
     if (!searchText || searchText.length == 0) {
       this.queryVariables.filter = null;
       return;
@@ -52,38 +52,13 @@ export class PublicationListComponent
     };
   }
 
-  private setQueryVariableFromRouteQueryParams(params: Params) {
-    const size = +params['size'];
-    const page = +params['page'];
-    const searchText = params['searchText'] as string | null;
-
-    if (size && size > 0) this.pageSize = this.queryVariables.take = size;
-    if (page && page > 0) this.queryVariables.skip = (page - 1) * size;
-    if (searchText && searchText.length > 0) this.setQueryFilter(searchText);
-  }
-
-  private setQueryParams() {
-    this.router.navigate([], {
-      relativeTo: this.activatedRoute,
-      queryParams: {
-        page:
-          this.queryVariables.skip == 0
-            ? 1
-            : this.queryVariables.skip / this.queryVariables.take + 1,
-        size: this.queryVariables.take,
-        searchText: this.searchText,
-      },
-      queryParamsHandling: 'merge',
-    });
-  }
-
   constructor(
     gql: GetPublicationsGQL,
-    private activatedRoute: ActivatedRoute,
-    private router: Router,
+    activatedRoute: ActivatedRoute,
+    router: Router,
     private alertService: MessageAlertService
   ) {
-    super(gql, {
+    super(gql, activatedRoute, router, {
       skip: 0,
       take: 20,
       sortBy: {
@@ -92,30 +67,7 @@ export class PublicationListComponent
     });
   }
 
-  searchText: string | null = null;
-  pageSizes: number[] = [10, 20, 50, 100];
-  pageSize: number = this.pageSizes[0];
-
   ngOnInit(): void {
-    this.activatedRoute.queryParams.subscribe({
-      next: (params) => {
-        this.setQueryVariableFromRouteQueryParams(params);
-        this.setQueryParams();
-        this.queryRef.refetch(this.queryVariables);
-      },
-    });
-  }
-
-  changePage(pageNumber: number) {
-    this.queryVariables.skip = (pageNumber - 1) * this.queryVariables.take;
-    this.queryRef.refetch(this.queryVariables);
-    this.setQueryParams();
-  }
-
-  changePageSize(selectedPageSize: number) {
-    this.queryVariables.take = selectedPageSize;
-    this.queryVariables.skip = 0;
-    this.queryRef.refetch(this.queryVariables);
-    this.setQueryParams();
+    this.init();
   }
 }
