+++ src/Web/kathanika-app/src/app/shared/types/pagination-query-variables.ts
@@ -1,0 +1,6 @@
+import { OperationVariables } from "@apollo/client/core";
+
+export interface PaginationQueryVariables extends OperationVariables {
+  skip: number,
+  take: number
+}
